Align flow steering commands with other objects
This patch aligns ibv_cmd_create/destroy_flow with creation/destruction
of other objects as of QP/CQ/SRQ,etc.

The above APIs expect to issue the commands with the kernel but not to
make the allocation/destruction of ibv_flow. This is the logic for other
objects around.

This change enables in the following patch some specific mlx4 handling
around cleanup on fatal in the destroy flow.

Note:
The change is done in one patch which involves verbs and providers as done
in the kernel in such cases when an internal API is changed.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>