rsockets: Periodically wake-up threads from rpoll
The changes to rpoll to use a signaling fd to wake up blocked threads,
combined with suspending polling while rsockets states may be changing
_should_ prevent any threads from blocking indefinitely in rpoll()
when a desired state change occurs.

However, to be extra safe and proactive that we don't end up hanging
an application, we periodically wake up any polling thread, so that it
can recheck its rsocket states.  This is simply a proactive measure to
allow apps to make forward progress and not in response to any known
issues.

The sleeping interval is set to an arbitrary value of 5 seconds, but
can be overridden using config files.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>