ibdiags: Do not use a post increment in a macro context
Usually we do not expect the htobe/etc implementations to be macros that
can multi-evaluate their argument, however, glibc and clang-7 get confused
and throw a warning as the __builtin_bswap64 is not being used.

There is no real reason for this sketchy construction, so just make it two
lines.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>