verbs: Introduce new verb to export IB objects
First step in sharing an IB object is to export it from one context to a
new one.

A new IB verb is introduced to export the IB from context to fd which
represents the context to export the object to.
The new object is assigned a new 32 bit handle which is returned to the
caller to be used later when sharing the object.

Exporting an IB object does not duplicate the object but instead create
new reference to it. The kernel keeps track of the reference counting so
when one process destroys an exported object (ex ibv_dealloc_pd) the
reference is updated and when reached to zero - the object is destroyed.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>
Signed-off-by: Yuval Shaia <yuval.shaia@oracle.com>