verbs: add export_to_fd verb
Export to FD verb allow exporting of IB HW object (ib_pd, ib_mr, etc..)
from the current context to another context identified by it's file
descriptor. The IB HW object refcnt is incremented every time the object
is exported. The object is assigned new 32 bit handle in the new
context and the handle is returned to the caller of this verb.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>